alerts-server: Refactor sql evolutions to add currencies table
The term market was used incorrectly were exchange should be used,
this is the initial commit for a big code refactor to fix the
mistake.

This change also abstracts alert related tables to link them to
the currencies table instead of keeping the whole currency data
in each alert table.

THE BUILD WILL SUCCEED AFTER COMPLETING THE BIG CODE REFACTOR